fix(trends-diff): Client/Server境界違反による dns Module not found を解消
`/trends/diff` のビルドで `Module not found: Can't resolve 'dns'` が発生していた問題を解消。

## 原因
Client Component (`diff-content.tsx`) が barrel `@/lib/ai/diff-summary` 経由で
日付ユーティリティだけを import していたが、barrel が `diff-summary-service.ts`
（`lib/prisma` を直接 import）を re-export していたため、Prisma v7 + adapter-pg
経由で `pg` → Node.js `dns` 依存が Client バンドルに引き込まれていた。

`next.config.ts` の `dns:false` fallback は webpack hook 内に書かれており、
Next.js 16 の dev デフォルトである Turbopack では適用されないため顕在化した。

## 対応 (A + B 併用)

### A: pure な date util を独立ファイルに切り出し
- `lib/ai/diff-summary/iso-week.ts` を新規作成し
  `getISOWeek/getPreviousISOWeek/getNextISOWeek/getWeekDateRange` を移動
  （副作用 import なし、Client から安全に import 可能）
- `diff-summary-service.ts` は新ファイルから re-import + 後方互換のため re-export
- barrel `index.ts` は server-only シンボルと date util を物理的に分離した
  経路で re-export するよう整理
- Client Component は barrel ではなく `@/lib/ai/diff-summary/iso-week` から
  直接 import するよう変更

### B: 再発防止
- `lib/prisma.ts` 冒頭に `import 'server-only'` を追加。Client から（直接・
  間接ともに）import された場合、build 時に明確なエラーで早期失敗する

## 検証
- `npm run docker:build` 成功（`/trends/diff` 含む全ルートのコンパイル成功）
- 関連テスト全 PASS:
  - `__tests__/lib/ai/diff-summary/diff-summary-service.test.ts` 11/11
  - `__tests__/api/ai/diff-summary-security.test.ts` 2/2
  - `__tests__/api/ai/diff-summary-fallback.test.ts` 5/5
